+++ google/cloud/pubsub/google_cloud_cpp_pubsub.bzl
@@ -42,6 +42,7 @@ google_cloud_cpp_pubsub_hdrs = [
     "internal/publisher_stub.h",
     "internal/publisher_stub_factory.h",
     "internal/pull_ack_handler.h",
+    "internal/pull_lease_manager.h",
     "internal/rejects_with_ordering_key.h",
     "internal/schema_auth_decorator.h",
     "internal/schema_logging_decorator.h",
@@ -111,6 +112,7 @@ google_cloud_cpp_pubsub_srcs = [
     "internal/publisher_stub.cc",
     "internal/publisher_stub_factory.cc",
     "internal/pull_ack_handler.cc",
+    "internal/pull_lease_manager.cc",
     "internal/rejects_with_ordering_key.cc",
     "internal/schema_auth_decorator.cc",
     "internal/schema_logging_decorator.cc",
