+++ google/cloud/pubsub/internal/pull_ack_handler.h
@@ -33,6 +33,7 @@ namespace google {
 namespace cloud {
 namespace pubsub_internal {
 GOOGLE_CLOUD_CPP_INLINE_NAMESPACE_BEGIN
+class PullLeaseManager;
 
 /**
  * Implements a `pubsub::ExactlyOnceAckHandler` suitable for blocking pull
@@ -40,56 +41,25 @@ GOOGLE_CLOUD_CPP_INLINE_NAMESPACE_BEGIN
  *
  * This is an implementation detail, hidden from the application.
  */
-class PullAckHandler : public pubsub::ExactlyOnceAckHandler::Impl,
-                       public std::enable_shared_from_this<PullAckHandler> {
+class PullAckHandler : public pubsub::ExactlyOnceAckHandler::Impl {
  public:
-  using Clock = std::function<std::chrono::system_clock::time_point()>;
-
   PullAckHandler(CompletionQueue cq, std::weak_ptr<SubscriberStub> w,
                  Options options, pubsub::Subscription subscription,
-                 std::string ack_id, std::int32_t delivery_attempt,
-                 Clock clock = std::chrono::system_clock::now);
+                 std::string ack_id, std::int32_t delivery_attempt);
   ~PullAckHandler() override;
 
   future<Status> ack() override;
   future<Status> nack() override;
   std::int32_t delivery_attempt() const override;
 
-  void StartLeaseLoop();
-
-  // The following functions are used in testing.
-  std::chrono::system_clock::time_point lease_deadline() const {
-    return lease_deadline_;
-  }
-  std::chrono::system_clock::time_point current_lease() const {
-    return current_lease_;
-  }
-
-  std::chrono::milliseconds LeaseRefreshPeriod() const;
-
  private:
-  void OnLeaseTimer(Status const& timer_status);
-  void OnLeaseExtended(std::chrono::system_clock::time_point new_deadline,
-                       Status const& status);
-
   CompletionQueue cq_;
   std::weak_ptr<SubscriberStub> stub_;
   Options options_;
   pubsub::Subscription subscription_;
   std::string ack_id_;
   std::int32_t delivery_attempt_;
-  Clock clock_;
-  // The absolute deadline to complete processing the message.
-  // The application can configure this value using
-  // `pubsub::MaxDeadlineTimeOption`
-  std::chrono::system_clock::time_point lease_deadline_;
-  // The duration for each lease extension.  Applications can configure this
-  // value using `pubsub::MaxDeadlineExtensionOption` and
-  // `pubsub::MinDeadlineExtensionOption`
-  std::chrono::seconds lease_extension_;
-  std::chrono::system_clock::time_point current_lease_;
-  std::mutex mu_;
-  future<void> timer_;
+  std::shared_ptr<PullLeaseManager> lease_manager_;
 };
 
 GOOGLE_CLOUD_CPP_INLINE_NAMESPACE_END
