+++ google/cloud/pubsub/pubsub_client_unit_tests.bzl
@@ -29,6 +29,7 @@ pubsub_client_unit_tests = [
     "internal/flow_controlled_publisher_connection_test.cc",
     "internal/ordering_key_publisher_connection_test.cc",
     "internal/pull_ack_handler_test.cc",
+    "internal/pull_lease_manager_test.cc",
     "internal/rejects_with_ordering_key_test.cc",
     "internal/sequential_batch_sink_test.cc",
     "internal/session_shutdown_manager_test.cc",
